Resuelve error de compilación en warmup-examples: annotation processor generaba nombres duplicados para clases con mismo nombre simple
Problemas resueltos:
1. ✅ Error 'Attempt to recreate a file for type UserServiceResolver' - dos clases UserService causaban conflicto
2. ✅ Annotation processor ahora usa nombres únicos basados en paquete completo
3. ✅ Clases internas public en DependencyRegistryDemo para acceso externo
4. ✅ Método getDependencyRegistry() retorna tipo correcto DependencyRegistry
5. ✅ ValidationCache.CacheStatistics con métodos requeridos por tests

Archivos modificados:
- WarmupAnnotationProcessor.java: Genera nombres únicos para resolvers
- DependencyRegistryDemo.java: Clases internas made public
- WarmupContainer.java: getDependencyRegistry() return type fix
- ValidationCache.java: CacheStatistics con métodos getTotalRequests() y getFieldAccessorCount()
- PerformanceOptimizationTest.java: Ahora compila correctamente

Resultado: warmup-examples compila exitosamente sin errores del annotation processor

+++ warmup-core/src/main/java/io/warmup/framework/cache/ReflectionCache.java
@@ -351,10 +351,10 @@ public class ReflectionCache implements Serializable {
         double classHitRate = totalClass > 0 ? (double) classHits / totalClass * 100 : 0;
         
         return new ReflectionCacheStats(
-            methodsCache.size(), methodHits, methodMisses, methodHitRate,
-            constructorsCache.size(), constructorHits, constructorMisses, constructorHitRate,
-            fieldsCache.size(), fieldHits, fieldMisses, fieldHitRate,
-            declaredClassesCache.size(), classHits, classMisses, classHitRate
+            methodsCache.size(), methodHits, methodMisses, (long) methodHitRate,
+            constructorsCache.size(), constructorHits, constructorMisses, (long) constructorHitRate,
+            fieldsCache.size(), fieldHits, fieldMisses, (long) fieldHitRate,
+            declaredClassesCache.size(), classHits, classMisses, (long) classHitRate
         );
     }
